systemtests/requirements.txt
PyYAML==5.1.2
requests==2.22.0
pytest-watch==4.2.0
pytest-bdd==3.2.1
